ItWorks

+++ libft/ft_memalloc.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memalloc.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mvaskiv <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: mvaskiv <mvaskiv@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/14 15:53:29 by mvaskiv           #+#    #+#             */
-/*   Updated: 2017/11/14 15:53:43 by mvaskiv          ###   ########.fr       */
+/*   Updated: 2017/11/17 17:45:16 by mvaskiv          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void		*ft_memalloc(unsigned int size)
+void		*ft_memalloc(size_t size)
 {
 	void			*str;
 	unsigned int	i;
