+++ libft/ft_memcpy.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memcpy.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mvaskiv <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: mvaskiv <mvaskiv@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/14 15:53:51 by mvaskiv           #+#    #+#             */
-/*   Updated: 2017/11/14 15:54:09 by mvaskiv          ###   ########.fr       */
+/*   Updated: 2017/11/17 18:38:05 by mvaskiv          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void		*ft_memcpy(void *dest, const void *src, unsigned int n)
+void		*ft_memcpy(void *dest, const void *src, size_t n)
 {
 	unsigned int	i;
 	char			*tmp_dest;
